shift_scheduler/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2
pydantic
httpx
pyjwt
python-dotenv
openpyxl
apscheduler
